--- react/__mocks__/vtex.store-components.js
@@ -1,23 +1,0 @@
-import React, { Component } from 'react'
-
-/**
- * ProductPrice Mocked Component.
- */
-export class ProductPrice extends Component {
-
-  static schema = {properties: {}}
-
-  render() {
-    return (
-      <div>ProductPrice</div>
-    )
-  }
-}
-
-export class ProductName extends Component {
-  render() {
-    return (
-      <div>ProductName</div>
-    )
-  }
-}
