+++ react/__mocks__/vtex.store-components/index.js
@@ -1,0 +1,27 @@
+import React, { Component } from 'react'
+
+export const DiscountBadge = ({ children }) => {
+  return <div className="discount-badge-mock">{children}</div>
+}
+
+const CollectionBadges = ({ children }) => {
+  return <div className="collection-badges-mock">{children}</div>
+}
+
+export class ProductPrice extends Component {
+  static schema = {properties: {}}
+
+  render() {
+    return (
+      <div>ProductPrice</div>
+    )
+  }
+}
+
+export class ProductName extends Component {
+  render() {
+    return (
+      <div>ProductName</div>
+    )
+  }
+}
